set POST /api/drink and POST /api/drink/<id>/img to be authorized endpoints
