+++ routes/health.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
-from .route_util import create_response
+from .route_util import create_response, is_authorized
 
 health_api = Blueprint('health', __name__, url_prefix = '/api')
