+++ routes/route_util.py
@@ -1,8 +1,11 @@
 import os
-import time
 
-from flask import current_app, jsonify
+from datetime import datetime, timedelta
+from flask import current_app, jsonify, session
 from werkzeug.utils import secure_filename
+
+from models.user import User
+from services.db import get_db, close_db
 
 def create_response(status, desc = "", data = []):
     return jsonify({ 'desc' : desc, 'data': data }), status
@@ -27,6 +30,53 @@ def save_image(img):
         return filename
     else:
         return ''
-    
-def current_milli_time():
-    return round(time.time() * 1000)
+
+#checks if the current session is authorized. deletes if the session is expired.
+def is_authorized():
+    if 'username' in session:
+        session_name = session['username']
+        c = get_db()
+        cursor = c.cursor()
+        query = User.fetch_session(session_name)
+        cursor.execute(query)
+        result = cursor.fetchone()
+
+        if result is None:
+            close_db(c)
+            return False
+        else:
+            if session_is_active(result['c_login_time']):
+                close_db(c)
+                return True
+            else:
+                delete_session(session_name, c)
+                return False
+
+def session_is_active(login_time):
+    login_duration_env = current_app.config['DB']['session_timeout']
+    login_duration = datetime.strptime(login_duration_env, '%H:%M:%S')
+    logout_time = login_time + timedelta(hours = login_duration.hour)
+    current_time = datetime.now()
+    return logout_time > current_time
+
+#deletes an active session. assumed to be last operation, closes connection
+def delete_session(session_name, c):
+    cursor = c.cursor()
+    query = User.fetch_session(session_name)
+    cursor.execute(query)
+    result = cursor.fetchone()
+
+    #no session found, no operation
+    if result is None:
+        close_db(c)
+        return False
+    else:
+        query = User.delete_session(result['c_id'])
+        cursor.execute(query)
+        c.commit()
+        close_db(c)
+
+        session.pop('username', None)
+
+        return True
+
